feat: add cluster functionality
Add code to track and update instance status as well as messages to
join and leave the cluster. Join and leave messages contain both the
joining/leaving UUID as well as the sending instance. The UUID of the
joining and leaving cluster should be the same, but we separate these
fields for now.

We keep the cluster servers as a sorted array of instances with UUID, last
seen time, and status. The array is sorted on the UUID.

Servers can be alive, suspected dead, or declared dead. We reserve one
state for unknown status, mostly as a precaution.

An added instance is automatically marked alive and added to the array.

When leaving the cluster, the instance is just marked as dead and it will
be cleaned up later.

Both joining and leaving the cluster will send a ack response, which can
be used to update the status of other instances since these messages
include gossip information.