feat: add cluster management functions
Add cluster functions `swim_cluster_join` and `swim_cluster_leave` that
allow you to join and leave an existing cluster. It also updates the
swim server code to optionally join an existing cluster by allowing the
address of a instance in the cluster to be passed as arguments to the
function.