feat: add support for indirect pings
If a node is not responding to a ping, we send an indirect ping request
to alive nodes in the view and ask them to ping the node.

We add a PING_REQ event and extend the ACK event with a dedicated field
that tells what node it is an acknowledge for and update the state
change functionality to deal with suspected and dead nodes properly.